requirements.txt
Streamlit
python-dotenv
google-generativeai